fixed the issue otp expiring early and refactored the code and routes used tiimeout() now ot delete otp and file have been moved to controller and routes
