+++ .env
@@ -1,0 +1,4 @@
+MONGO_URL='mongodb+srv://Daman:Ghattu12@cluster0.xtzogzt.mongodb.net/'
+MAIL_USER=sdaman830@gmail.com
+MAIL_PASS=mrxlapciyzljlfvk
+MAIL_HOST='smtp.gmail.com'
