+++ server/src/modules/usermanagement/privateroute.js
@@ -1,16 +1,17 @@
 const express = require("express");
 const router = express.Router();
-const jwt = require('jsonwebtoken');
-const getUserDetails = require('./dto');
-const jwtSecret =   "ad8cfdfe03c3076a4acb369ec18fbfc26b28bc78577b64da02646cd7bd0fe9c7d97cab"; // Replace with your actual secret
+const jwt = require("jsonwebtoken");
+const getUserDetails = require("./controllers/dto");
+const jwtSecret =
+  "ad8cfdfe03c3076a4acb369ec18fbfc26b28bc78577b64da02646cd7bd0fe9c7d97cab"; // Replace with your actual secret
 
 // Middleware to protect the route and verify the token
 function protectRoute(req, res, next) {
-  const token = req.cookies.jwt; 
+  const token = req.cookies.jwt;
   // console.log(token)
 
   if (!token) {
-    return res.status(401).json({ message: 'Unauthorized' });
+    return res.status(401).json({ message: "Unauthorized" });
   }
 
   try {
@@ -19,17 +20,17 @@ function protectRoute(req, res, next) {
 
     // The token is valid, and 'decoded' contains user information
     const userId = decoded.id;
-    
-   // Attach the user details to the 'req' object
-   req.user = {
-    id: userId,
-    // other user details...
-  };
 
-  // Allow the request to proceed
-  next();
+    // Attach the user details to the 'req' object
+    req.user = {
+      id: userId,
+      // other user details...
+    };
+
+    // Allow the request to proceed
+    next();
   } catch (err) {
-    return res.status(401).json({ message: 'Unauthorized' });
+    return res.status(401).json({ message: "Unauthorized" });
   }
 }
